backend/requirements.txt
# Django
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
django-filter>=24.0,<25.0
django-extensions>=3.2,<4.0
django-storages[boto3]>=1.14,<2.0

# Database
psycopg[binary]>=3.2,<4.0
dj-database-url>=2.2,<3.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0

# Async / WebSockets
channels>=4.1,<5.0
channels-redis>=4.2,<5.0
daphne>=4.1,<5.0

# Task queue
celery>=5.4,<6.0
redis>=5.0,<6.0

# Storage
boto3>=1.35,<2.0

# Media processing
Pillow>=10.4,<11.0

# Server
gunicorn>=22.0,<23.0
uvicorn[standard]>=0.30,<1.0

# LiveKit
livekit-api>=0.7,<1.0

# Utilities
python-decouple>=3.8,<4.0
python-dateutil>=2.9,<3.0

# Dev & Testing
ruff>=0.6,<1.0
pytest>=8.3,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=5.0,<6.0
factory-boy>=3.3,<4.0
